aiogptbot/requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
redis==5.0.4
openai==1.30.1
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.5.2
